Add files via upload
Download the dataset named churn_prediction.csv and also download the ipynb file . Open jupyter notebook and upload the code to the notebook and click run all. Make sure that you have changed the path where the csv file is saved. This model can able to predict the churns with accuracy of 85%.